requirements.txt
pinterest-dl
python-dotenv
fpdf2
flask
b2sdk